#27 popup router interface & change component name
