+++ src/components/Navigation.tsx
@@ -11,7 +11,6 @@ const Navigation = (props: any) => {
       id: '1',
       name: t('app.navigation.product'),
       onClick() {
-        // navigate('/movie?category=popular')
         navigate('/overall/movie/popular')
       },
     },
