+++ src/components/section/styles/ToggleBar.scss
