+++ src/styles/DetailPage.scss
@@ -31,10 +31,8 @@
           width: 100%;
           height: calc(100vw / 2.22);
           padding: 0;
-          // border-radius: 8px;
           .image-background {
             width: 100%;
-            // padding: 15px;
             height: 100%;
             position: relative;
             box-sizing: border-box;
