+++ src/styles/MoviePage.scss
@@ -5,7 +5,7 @@
   display: grid;
   grid-template-columns: 300px 2.5fr;
   .mobile & {
-    @include flexColumn();
+    @include flexColumn;
   }
   & > .filter-total-container {
     position: relative;
@@ -15,10 +15,7 @@
     }
   }
   & > .movie-item-container {
-    display: flex;
-    flex-direction: column;
-    padding: 16px;
-    gap: 16px;
+    @include flexColumnCenterPaddingGap(16px, 16px);
 
     & > .item-grid {
       display: grid;
