+++ src/types/popup/RouterTypes.ts
@@ -1,8 +1,8 @@
 export interface RegisterRouter {
   name: string
-  title? : string
-  maxProgress? : number
-  progress? : number
+  title?: string
+  maxProgress?: number
+  progress?: number
   component: any
 }
 export interface RouterPushParams {
@@ -14,9 +14,13 @@ export interface RouterPushParams {
 export interface RouteHistory {
   name: string
   component: any
-  title : string
-  progress? : number
-  maxProgress? : number
+  title: string
+  progress?: number
+  maxProgress?: number
   props?: Record<string, any>
   events?: Record<string, any>
+}
+export interface PopupComponentProps {
+  close: () => void
+  push: (route: RouterPushParams) => void
 }
