+++ src/views/signup_popup/SignupResult.tsx
@@ -1,9 +1,7 @@
-import { RouterPushParams } from '@/types/popup/RouterTypes'
+import { PopupComponentProps, RouterPushParams } from '@/types/popup/RouterTypes'
 import React from 'react'
 
-interface RouterProps {
-  close: () => void
-  push: (route: RouterPushParams) => void
+interface RouterProps extends PopupComponentProps {
   userInfo: { [key: string]: string }
 }
 const SignupResult = (props: RouterProps) => {
